Fix size of SPARC64 fprs register
Some of the code was treating it as 32-bit, some as 64-bit.  When qemu
is run on a little-endian host that sort of works, the relevant 32-bits
just happens to be in the right place.  On a big-endian host that
places the relevant bits in the wrong place, the register contains
zeros.